Adds REDIS root handling to ticket management
All ticket operations now happends below a config file specified root
($CONFIG['REDIS_ROOT']).  It is advised to end the choose root with '.',
so keys stay readable and collisions are avoided.

Documentation for twitter client config has been added. The same warning
applies to the twitter Redis key root. The client is modified so it does
*NOT* append a dot between root and the requested key.